+++ packages/app/playwright/commons/seedWallet.ts
@@ -4,6 +4,8 @@ import { Address, BaseAssetId, Wallet } from 'fuels';
 
 import { getAccount, ALT_ASSET } from '../mocks';
 
+import { getGasConfig } from './gas';
+
 const { GENESIS_SECRET, VITE_FUEL_PROVIDER_URL } = process.env;
 
 export async function seedCurretAccount(page: Page, amount: BN) {
@@ -16,18 +18,19 @@ export async function seedWallet(address: string, amount: BN) {
     GENESIS_SECRET,
     VITE_FUEL_PROVIDER_URL
   );
+  const { gasPrice, gasLimit } = await getGasConfig(genesisWallet.provider);
   const transfETH = await genesisWallet.transfer(
     Address.fromString(address),
     amount,
     BaseAssetId,
-    { gasPrice: 1 }
+    { gasPrice, gasLimit }
   );
   await transfETH.wait();
   const transfAsset = await genesisWallet.transfer(
     Address.fromString(address),
     amount,
     ALT_ASSET.assetId,
-    { gasPrice: 1 }
+    { gasPrice, gasLimit }
   );
   await transfAsset.wait();
 }