+++ packages/app/src/systems/Transaction/pages/ViewActivity/ViewActivity.tsx
@@ -3,7 +3,7 @@ import { Address } from 'fuels';
 import { useNavigate } from 'react-router-dom';
 
 import { ActivityList } from '../../components/ActivityList/ActivityList';
-import { useTxs } from '../../hooks/useTxs';
+import { useTransactionHistory } from '../../hooks';
 
 import { useAccounts } from '~/systems/Account';
 import { Layout, Pages } from '~/systems/Core';
@@ -13,24 +13,24 @@ export function ViewActivity() {
   const navigate = useNavigate();
   const networks = useNetworks();
   const providerUrl = networks?.selectedNetwork?.url;
-  const { account, isLoading } = useAccounts();
+  const { account, isLoading: isLoadingAccounts } = useAccounts();
 
   const address = account
     ? Address.fromAddressOrString(account?.address).toB256()
     : '';
-  const { isLoadingTx, txs } = useTxs({
+  const { isLoading: isLoadingTx, transactionHistory } = useTransactionHistory({
     address,
     providerUrl,
   });
 
   return (
-    <Layout title="History" isLoading={isLoading}>
+    <Layout title="History" isLoading={isLoadingTx || isLoadingAccounts}>
       <Layout.TopBar onBack={() => navigate(Pages.wallet())} />
       <Layout.Content>
         <Box.Stack gap="$4">
           <ActivityList
-            txs={txs ?? []}
-            isLoading={isLoadingTx || isLoading || !account}
+            txs={transactionHistory ?? []}
+            isLoading={isLoadingTx || isLoadingAccounts || !account}
             ownerAddress={address}
           />
         </Box.Stack>